create upgrade-to-2020-conventions branch
A lot has changed since revista was released as a sample app in 2018, in
both Elixir and Terraform. Mix releases no longer need distiller as an
explicit dependency, and they work a bit differently. Plus, the
Gruntwork file layout really isn't necessary with Terraform workspaces.

This branch is mostly about updating Elixir, Terraform, and the
deployment process, but I'll go ahead and upgrade all of the
dependencies while I'm at it.